mark APQ and query plan caches as license key functionality
