+++ apollo-router/src/configuration/mod.rs
@@ -374,6 +374,7 @@ impl Configuration {
                 },
             );
         }
+
         Ok(self)
     }
 }
